Move operator-mirroring tests into UltimateNormalFormTest
Uses runUnfTest (UnfTransformer directly) instead of
SimplificationTest.runSimplificationTest, matching the workflow already
used for the other UnfTransformer-level tests in this class.
